july-2021-leetcoding-challenge/src/test/java/org/redquark/leetcode/challenge/Problem24_WordLadderIITest.java
package org.redquark.leetcode.challenge;

import org.junit.jupiter.api.Test;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.Collections;
import java.util.List;

import static org.junit.jupiter.api.Assertions.assertEquals;

class Problem24_WordLadderIITest {

    private final Problem24_WordLadderII testObject = new Problem24_WordLadderII();

    @Test
    void testFindLadders() {
        String beginWord = "hit";
        String endWord = "cog";
        List<String> wordList = Arrays.asList("hot", "dot", "dog", "lot", "log", "cog");
        List<List<String>> expected = new ArrayList<>();
        expected.add(Arrays.asList("hit", "hot", "dot", "dog", "cog"));
        expected.add(Arrays.asList("hit", "hot", "lot", "log", "cog"));
        assertEquals(expected, testObject.findLadders(beginWord, endWord, wordList));

        beginWord = "hit";
        endWord = "cog";
        wordList = Arrays.asList("hot", "dot", "dog", "lot", "log");
        expected = Collections.emptyList();
        assertEquals(expected, testObject.findLadders(beginWord, endWord, wordList));
    }
}